backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI
openai==1.10.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# APScheduler
apscheduler==3.10.4

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3

# LangGraph (for agent orchestration)
langgraph==0.0.20
langchain==0.1.4
langchain-core==0.1.16

# HTTP client
httpx==0.26.0

# Environment variables
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7
